Разделил написанный код на отдельные модули
